% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R62S3.R
\name{R62S3}
\alias{R62S3}
\title{S3 Method Generator from R6 Class}
\usage{
R62S3(R6Class, dispatchClasses = list(R6Class), assignEnvir = parent.env(environment()))
}
\arguments{
\item{R6Class}{R6ClassGenerator to generate public methods from}

\item{dispatchClasses}{list of classes to assign S3 dispatch methods on}

\item{assignEnvir}{environment in which to assign the S3 generics/methods, default Global Environment}
}
\value{
Assigns methods and generics to the chosen environment.
}
\description{
Auto-generates S3 generics and public methods from an R6 Class.
}
\details{
Searches in a given R6 class for all public methods that are not 'initialize' or 'clone'.
For each method if a generic does not already exist, one is created and assigned to the given environment.
Methods are created for every generic, following standard S3 convention.
}
\examples{
printMachine <- R6::R6Class("printMachine",
public = list(initialize = function() {},
printer = function(str) {print(str)}))
pm <- printMachine$new()
R62S3(printMachine, assignEnvir = .GlobalEnv)
pm$printer("Test String A")
printer(pm, "Test String B")

}
