% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R62S3.R
\name{R62S3}
\alias{R62S3}
\title{S3 Method Generator from R6 Class}
\usage{
R62S3(R6Class, dispatchClasses, assignEnvir)
}
\arguments{
\item{R6Class}{The R6ClassGenerator or Classname to generate public methods from}

\item{dispatchClasses}{a list of classes to assign S3 dispatch methods on}

\item{assignEnvir}{the environment in which to assign the S3 generics/methods, default Global Environment}
}
\value{
Assigns methods and generics to the chosen environment.
}
\description{
Auto-generates S3 generics and public methods from an R6 Class.
}
\details{
The input must either be of class R6ClassGenerator or a character
  string naming the R6ClassGenerator. Also assumes the classname is the same
  as the R6ClassGenerator name.
}
\examples{
printMachine <- R6::R6Class("printMachine",
public = list(initialize = function() {},
printer = function(str) {print(str)}))
pm <- printMachine$new()
R62S3(printMachine)
pm$printer("Test String A")
printer(pm, "Test String B")

}
